Day 7, part 1 + 2 completed.
Added a small excursion on Clojures queues to day 6, maybe I can put
this in a different file later on.

Via recommendation, switched to the `bb-clerk` branch for Clerk, which
allows me to run Babashka tests without needing conditionals on
Clerk-specific code. I’ll have to see if I can also run the notebooks in
Babashka.